refactor: rename format component
